perf(lxl-web): Compute qualiferSuggestions once
took ~14% cpu when profiling locally with stress_search2_api_simple.py
